Drag and Drop available, new adapter
