+++ .idea/libraries/DynamicGrid_f81c028535.xml
@@ -1,0 +1,10 @@
+<component name="libraryTable">
+  <library name="DynamicGrid-f81c028535">
+    <CLASSES>
+      <root url="jar://$PROJECT_DIR$/app/build/intermediates/exploded-aar/com.github.askerov/DynamicGrid/f81c028535/jars/classes.jar!/" />
+      <root url="file://$PROJECT_DIR$/app/build/intermediates/exploded-aar/com.github.askerov/DynamicGrid/f81c028535/res" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES />
+  </library>
+</component>
