+++ app/build.gradle
@@ -50,4 +50,6 @@ dependencies {
     compile 'com.android.support:cardview-v7:23.4.0'
     compile 'com.android.support:recyclerview-v7:23.4.0'
     compile 'io.github.kobakei:ratethisapp:1.0.5'
+    compile 'com.github.askerov:DynamicGrid:f81c028535'
+
 }
