+++ app/src/main/java/com/pressurelabs/flow/FlowElement.java
@@ -3,6 +3,8 @@ package com.pressurelabs.flow;
 import android.os.Parcel;
 import android.os.Parcelable;
 
+import java.util.UUID;
+
 /**
  * Flow_V2
  *
@@ -21,6 +23,12 @@ public class FlowElement implements Parcelable{
 
     private String timeUnits;
 
+    public String getUuid() {
+        return uuid;
+    }
+
+    private String uuid;
+
     private int location;
 
     public FlowElement(String name, int timeEst, String units) {
@@ -28,6 +36,7 @@ public class FlowElement implements Parcelable{
         this.elementName = name;
         this.timeEstimate = timeEst;
         this.timeUnits = units;
+        this.uuid = UUID.randomUUID().toString();
     } //End of constructor
 
     /*~~~~~ Getter and Setter Methods: ~~~~~*/
