+++ app/src/main/res/layout/sb_drag_drop_gridview.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
+<org.askerov.dynamicgrid.DynamicGridView
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/drag_drop_gridview"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:paddingTop="?android:attr/actionBarSize"
+    android:clipToPadding="false"
+    android:columnWidth="90dp"
+    android:numColumns="auto_fit"
+    android:verticalSpacing="25dp"
+    android:horizontalSpacing="10dp"
+    android:stretchMode="columnWidth"
+    android:gravity="center"
+    android:background="?android:attr/activatedBackgroundIndicator">
+</org.askerov.dynamicgrid.DynamicGridView>
