+++ app/src/main/res/menu/menu_sandbox.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+    <item
+        android:id="@+id/action_reorder_elements"
+        android:title="@string/action_reorder_elements"
+        android:icon="@drawable/reorder_white_36dp"
+        app:showAsAction="always"/>
+
+
+    <item
+        android:id="@+id/action_flow_statistics"
+        android:icon="@drawable/times_completed_white_36dp"
+        android:title="@string/action_flow_statistics"
+        app:showAsAction="always"/>
+
+</menu>
