+++ app/src/main/res/menu/menu_the_hub.xml
@@ -10,7 +10,7 @@
         compat:showAsAction="always"/>
 
     <item android:id="@+id/action_delete_flows"
-        android:icon="@drawable/delete_all_white_24dp"
+        android:icon="@drawable/delete_all_white_36dp"
         android:title="@string/action_delete_flow"
         compat:showAsAction="always"/>
