+++ app/src/main/res/values/styles.xml
@@ -19,7 +19,8 @@
 
     <style name="AppTheme" parent="Theme.AppCompat.Light.NoActionBar">
         <item name="android:statusBarColor">@color/abc_secondary_text_material_light</item>
-        <item name="android:windowNoTitle">true</item>
+        <item name="colorPrimary">@color/primary_dark_material_light</item>
+        <item name="colorAccent">@color/colorPrimaryDark</item>
     </style>
 
     <style name="AppTheme.NoActionBar">
